# src/swagger.yaml
swagger: '2.0'
info:
  description: >-
    Facilitar a integração social de pessoas através de exercícios físicos
    feitos coletivamente
  version: 1.0.0
  title: Honoo
  termsOfService: 'http://honoo.io/terms/'
  contact:
    email: honoo@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'b5els.sse.codesandbox.io'
basePath: /
...
schemes:
  - https
paths:
  /user:
    post:
      tags:
        - user
      summary: Cadastrar um novo usuario
      description: ''
      operationId: addUser
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Necessário um objeto Usuario
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '201':
          description: Usuario Criado
          schema:
            $ref: '#/definitions/User'
        '405':
          description: Entrada Invalida
      x-swagger-router-controller: User
    put:
      tags:
        - user
      summary: Atualizar um usuario
      description: ''
      operationId: updateUsuario
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: ''
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: Usuário Atualizado
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Invalid ID supplied
        '404':
          description: Usuario não encontrado
        '405':
          description: Validation exception
      x-swagger-router-controller: User
    get:
      tags:
        - user
      summary: Lista de Usuarios
      description: ''
      operationId: consultaUsuarios
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: Usuários Listados
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        '400':
          description: Nenhum usuario encontrado
        '405':
          description: Validation exception
      x-swagger-router-controller: User
  '/user/{userId}':
    get:
      tags:
        - user
      summary: Enconte um usuario pelo seu ID
      description: ''
      operationId: getUserById
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: Id do usuario que quer retorno
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Operacão Realizada com Sucesso
          schema:
            $ref: '#/definitions/User'
        '204':
          description: Não a conteudo para ser entregue
        '400':
          description: Invalid ID supplied
        '404':
          description: Usuario não Encontrado
      x-swagger-router-controller: User
    delete:
      tags:
        - user
      summary: Deletar Usuario
      description: ''
      operationId: deleteUser
      produces:
        - application/json
      parameters:
        - name: api_key
          in: header
          required: false
          type: string
        - name: userId
          in: path
          description: Id do Usuario que deseja deletar
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Usuario apagado com sucesso
        '400':
          description: Invalid ID supplied
        '404':
          description: Usuario não Encontrado
      x-swagger-router-controller: User
  /user/login:
    post:
      tags:
        - user
      summary: Logar usuario dentro do Sistema
      description: ''
      operationId: loginUser
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Necessário um objeto User
          required: true
          schema:
            $ref: '#/definitions/login'
      responses:
        '200':
          description: successful operation
          headers:
            X-Rate-Limit:
              type: integer
              format: int32
              description: calls per hour allowed by the user
            X-Expires-After:
              type: string
              format: date-time
              description: date in UTC when token expires
          schema:
            type: string
        '400':
          description: Invalid username/password supplied
      x-swagger-router-controller: User
...
definitions:
...
  login:
    type: object
    properties:
      login:
        type: string
      senha:
        type: string
...
  User:
    type: object
    required:
      - email
      - password
      - username
    properties:
      id:
        type: integer
        format: int64
      username:
        type: string
      email:
        type: string
      password:
        type: string
      birthdate:
        type: string
        format: date
      sex:
        type: string
    example:
      password: password
      birthdate: '2000-01-23'
      sex: sex
      id: 0
      email: email
      username: username
...